requirements.txt
# Web框架
fastapi>=0.123.9
uvicorn>=0.38.0

# 数据库
alembic==1.12.1
pymysql==1.1.0

# 认证和安全
PyJWT>=2.9.0
bcrypt>=4.2.0
python-multipart>=0.0.20

# 环境变量
python-dotenv>=1.2.1

# 数据验证
pydantic>=2.12.5
pydantic-settings==2.1.0

# HTTP客户端
requests>=2.32.0

# 图像处理
Pillow>=12.0.0

# 工具库
python-dateutil==2.8.2

# 日志
loguru==0.7.2

# CORS (FastAPI内置，无需单独安装)

 

